Make a unified sourcedstring module that works in Python 2 and 3
