+++ gate/sourcedstring3.py
